refactor : transformed every argument of functions in generic stats from datetime to strings
This refactor aims to standardised the way we pass dates to the
functions of the stats repository.
It also corrects some imprecisions related to the retrieval of data per
weeks, with a little caveat : when we retrieve the time per task for
week, we ignore the part of the week that isn't in the time span
specified, but in the year generic stats, the bar chart (without task
time) leaves behind the partial weeks.