fix : month generic stats
The way we fetched the data for the weeks was all wrong : not including
partial weeks, or including the task time for all the week instead of
just from the beginning date stated, which caused some problems for the
stats.
The bugg has been spotted because 2025 doesn't begin a Monday, contrary
to 2024, therefore sqlite gets week 1 as 0.