fix : year generic stats
Like in commit a24d824, a change in the way we calculate the week number
was needed.
To simplify, this commit also restrain to get all the weeks for only one
year at a time for the time being.

Add basic tests for generic stats. On refactoring, allows us to check
that the essence of what it should return is unchanged.