feat : adding route for generic stats
This commit adds a /stat/generic route that can take parameters in order
to group the different methods in only one and simplify things. The
separation was good while there was only a generic stat for the current
period, but it's better to modify things so there are not 4 or 5 changes
to do every single time in every generic stat route.