feat : time task ratio for past dates
This commit allows time task ratio to get data for past dates, which
completes the generic stats refactoring.
The function feels a little weird being parameterless, but it's only
because it always inherits from BaseStatsService which always receive at
least a period and completes the date if none is given.