fix : subtask not NULL when no subtask on new task
This commit fix the storing of subtask as empty string in sqlite which couldn't be retrieved with repository condition 'subtask IS NULL'.
It also fixes a problem related to the presence or not of flows in the config file (we'll probably store them in a json flow file which will necessitate a path to be given but that's it for now, not done.